Fixed Docker Entrypoint Isues, moved debug (text) driver in own folder. Fixed import from idt.h. Added multiboot base.
