+++ src/kernel/arch/i686/mb2/mb.c
@@ -1,0 +1,1 @@
+#include "mb.h"
