+++ src/kernel/arch/i686/mb2/mb.h
@@ -1,0 +1,2 @@
+#pragma once
+#define m 0
